OPE-949 remove startDate and endDate from operation index

+++ qa/migration-tests/src/test/java/org/camunda/operate/migration/ValidationTest.java
@@ -183,7 +183,7 @@ public class ValidationTest {
          (long)config.getWorkflowInstanceCount() - config.getCountOfCancelOperation());
      // Incidents resolved or Workflow instances are canceled
      assertThat(incidentCount).isBetween(
-        (long)config.getIncidentCount() - maxNotRunning,
+        config.getIncidentCount() - maxNotRunning,
         (long) config.getIncidentCount()
      );
   }
